+++ NEWS.md
@@ -1,4 +1,4 @@
-ParallelLogger 3.3.2
+ParallelLogger 3.4.0
 ====================
 
 Changes
@@ -6,6 +6,10 @@ Changes
 1. Exposing `getThreadNumber()` function that can be used to determine which thread we're in (including the main thread).
 
 2. Updating `selectFromList()` to support selection from R6 objects.
+
+3. Added `getPhysicalMemory()` function.
+
+4. Added `setAndromedaMemoryLimit` argument to `makeCluster()`. When TRUE, the andromedaMemoryLimit option will be set in each thread to be either the global andromedaMemoryLimit / numberOfThreads or 75% of the system memory / number of threads.
 
 
 ParallelLogger 3.3.1